+++ opensearchtransport/opensearchtransport.go
@@ -116,8 +116,6 @@ type Client struct {
 
 	includeDedicatedClusterManagers bool
 
-	includeDedicatedClusterManagers bool
-
 	compressRequestBody  bool
 	pooledGzipCompressor *gzipCompressor